pages/requirements.txt

absl-py==2.2.1
accelerate==1.4.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
asttokens==3.0.0
astunparse==1.6.3
attrs==25.1.0
beautifulsoup4==4.12.3
blinker==1.9.0
blis==1.2.0
bs4==0.0.2
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.20.0
cloudpickle==3.1.1
colorama==0.4.6
confection==0.1.5
contourpy==1.3.1
cryptography==44.0.2
curated-tokenizers==0.0.9
curated-transformers==0.1.1
cycler==0.12.1
cymem==2.0.11
dataclasses-json==0.5.14
datasets==3.3.2
decorator==5.2.1
dill==0.3.8
distro==1.9.0
dotenv==0.9.9
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
en_core_web_trf @ https://github.com/explosion/spacy-models/releases/download/en_core_web_trf-3.8.0/en_core_web_trf-3.8.0-py3-none-any.whl#sha256=272a31e9d8530d1e075351d30a462d7e80e31da23574f1b274e200f3fff35bf5
executing==2.2.0
faiss-cpu==1.10.0
Farama-Notifications==0.0.4
filelock==3.17.0
Flask==3.1.0
flatbuffers==25.2.10
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.12.0
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.163.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-genai==1.5.0
google-generativeai==0.8.4
google-pasta==0.2.0
googleapis-common-protos==1.69.1
greenlet==3.1.1
grpcio==1.71.0
grpcio-status==1.63.0rc1
gunicorn==20.0.4
gym==0.26.2
gym-notices==0.0.8
gymnasium==1.1.1
h11==0.14.0
h5py==3.13.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.29.1
idna==3.10
importlib_metadata==7.2.1
interchange==2021.0.4
ipython==9.0.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jsonpatch==1.33
jsonpickle==4.0.2
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keras==3.9.1
kiwisolver==1.4.8
langchain==0.3.20
langchain-core==0.3.45
langchain-text-splitters==0.3.6
langcodes==3.5.0
langsmith==0.3.13
language_data==1.3.0
libclang==18.1.1
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
ml_dtypes==0.5.1
mlxtend==0.23.4
monotonic==1.6
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
murmurhash==1.0.12
mypy-extensions==1.0.0
namex==0.0.8
narwhals==1.29.0
neo4j==5.28.1
networkx==3.4.2
nltk==3.9.1
numexpr==2.10.2
numpy==1.26.4
openai==0.28.0
openapi-schema-pydantic==1.2.4
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.14.1
orjson==3.10.15
packaging==23.2
pandas==2.2.3
pansi==2024.11.0
parso==0.8.4
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.4
pillow==10.4.0
plotly==6.0.1
preshed==3.0.9
prompt_toolkit==3.0.50
propcache==0.3.0
proto-plus==1.26.1
protobuf==4.25.6
psutil==7.0.0
pure_eval==0.2.3
py2neo==2021.2.4
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
PyGetWindow==0.0.9
Pygments==2.19.1
PyMsgBox==1.0.9
pyparsing==3.2.1
PyPDF2==3.0.1
pypdfium2==4.30.1
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytweening==1.2.0
pytz==2025.1
pyvis==0.3.2
PyYAML==6.0.2
qrcode==8.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.23.1
rsa==4.9
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==3.4.1
setuptools==75.8.2
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
spacy==3.8.4
spacy-curated-transformers==0.3.0
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.39
srsly==2.5.1
stack-data==0.6.3
streamlit==1.32.0
sympy==1.13.1
tenacity==8.5.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==3.0.0
thinc==8.3.4
threadpoolctl==3.5.0
tiktoken==0.9.0
tokenizers==0.21.0
toml==0.10.2
torch==2.6.0
torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.49.0
typer==0.15.2
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3.1
uritemplate==4.1.1
urllib3==2.3.0
validators==0.34.0
wasabi==1.1.3
watchdog==6.0.0
wcwidth==0.2.13
weasel==0.4.1
websockets==14.2
Werkzeug==3.1.3
wheel==0.45.1
wikipedia==1.4.0
wrapt==1.17.2
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
zstandard==0.23.0
